backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.6.1
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiosqlite==0.19.0
pytest==7.4.3
requests==2.31.0
Pillow==10.1.0 